uploading profiles and viewing is working

+++ Firmware/MaD_Firmware/include/Main/Communication/CRC.h
@@ -2,5 +2,5 @@
 #define CRC_H
 #include <stddef.h>
 #include <stdint.h>
-uint8_t crc8(uint8_t *addr, uint8_t len);
+uint8_t crc8(uint8_t *addr, uint16_t len);
 #endif
